feat(front): move markdown parsing to web worker
